backend/pyproject.toml
[project]
name = "ADK-agentic-scaffolding"
version = "0.0.1"
description = "ADK backend with FastAPI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "pydantic-settings>=2.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src"]